# feat : elasticsearch // #fix : login , KakaoPay

+++ server/src/main/java/com/main/project/order/controller/KakaoController.java
@@ -10,6 +10,7 @@ import com.main.project.member.repository.MemberRepository;
 import com.main.project.member.repository.RefreshTokenRepository;
 import com.main.project.member.service.MemberService;
 import com.main.project.order.dto.KakaoPayApprovalVO;
+import com.main.project.order.dto.KakaoPayReadyVO;
 import com.main.project.order.dto.OrderDto;
 import com.main.project.order.entity.Order;
 import com.main.project.order.entity.Orderproduct;
@@ -22,6 +23,8 @@ import com.main.project.product.entity.Product;
 import com.main.project.product.mapper.ProductMapper;
 import com.main.project.product.repository.ProductRepository;
 import com.main.project.product.service.ProductService;
+import com.main.project.search.document.Eproduct;
+import com.main.project.search.service.EproductService;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.extern.java.Log;
@@ -50,6 +53,8 @@ public class KakaoController {
     private final OrderproductRepository orderproductRepository;
     private final OrderproductService orderproductService;
     private final MemberRepository memberRepository;
+    private final EproductService eproductService;
+    private final ProductMapper productMapper;
 
     private final ProductRepository productRepository;
     @Setter(onMethod_ = @Autowired)
@@ -77,7 +82,7 @@ public class KakaoController {
         for(int i = 0; i < productList.size(); i++){
             sum += productList.get(i).getprice();
         }
-        Order createdorder = orderService.createOrder(order,member,sum);
+        Order createdorder = orderService.createOrder(order,member,sum,Long.valueOf(0));
         return ResponseEntity.ok(kakaopay.kakaoPayReadybucket(productList, createdorder));
     }
 
@@ -95,7 +100,7 @@ public class KakaoController {
         for(int i = 0; i < productList.size(); i++){
             sum += productList.get(i).getPrice();
         }
-        Order createdorder = orderService.createOrder(order,member,sum);
+        Order createdorder = orderService.createOrder(order,member,sum, product.getProductId());
 
         return ResponseEntity.ok(kakaopay.kakaoPayReady(productList, createdorder));
 
@@ -107,20 +112,39 @@ public class KakaoController {
         log.info("kakaoPaySuccess pg_token : " + pg_token);
         KakaoPayApprovalVO kakaoPayApprovalVO = kakaopay.kakaoPayInfo(pg_token);
         Member member = memberService.findMember(kakaoPayApprovalVO.getPartner_user_id());
-        List<queryget.findbypid> productIdList = orderproductRepository.findAllByMemberId(member.getMemberId());
-        productIdList.forEach(productId -> {
-            Product product = productService.findProduct(productId.getproduct_id());
+        if(kakaoPayApprovalVO.getItem_code().equals("single")){
+            // change product state
+            Product product = productService.findProduct(Long.valueOf(kakaoPayApprovalVO.getPartner_order_id()));
             product.setIssell(true);
+            Eproduct eproduct = productMapper.productToEproduct(product);
+            eproduct.setSell("sold");
+            eproductService.addEproduct(eproduct);
             productRepository.save(product);
+            // sava point
             Member seller = memberService.findVerifiedMember(product.getMember().getMemberId());
             seller.setMoney(seller.getMoney() + product.getPrice()*10/100);
             memberRepository.save(seller);
-            List<queryget.findbyorderpid> orderproductIdList = orderproductService.findOrderproductdelete(member.getMemberId(),product.getProductId());
-            orderproductIdList.forEach(orderproductId -> {
-                Orderproduct orderproduct = orderproductService.findorderproduct(orderproductId.getorderproduct_id());
-                orderproductRepository.delete(orderproduct);
+        }
+        else{
+            List<queryget.findbypid> productIdList = orderproductRepository.findAllByMemberId(member.getMemberId());
+            productIdList.forEach(productId -> {
+                Product product = productService.findProduct(productId.getproduct_id());
+                product.setIssell(true);
+                Eproduct eproduct = productMapper.productToEproduct(product);
+                eproduct.setSell("sold");
+                eproductService.addEproduct(eproduct);
+                productRepository.save(product);
+                Member seller = memberService.findVerifiedMember(product.getMember().getMemberId());
+                seller.setMoney(seller.getMoney() + product.getPrice()*10/100);
+                memberRepository.save(seller);
+                List<queryget.findbyorderpid> orderproductIdList = orderproductService.findOrderproductdelete(member.getMemberId(),product.getProductId());
+                orderproductIdList.forEach(orderproductId -> {
+                    Orderproduct orderproduct = orderproductService.findorderproduct(orderproductId.getorderproduct_id());
+                    orderproductRepository.delete(orderproduct);
+                });
             });
-        });
+        }
+
         model.addAttribute("info", kakaoPayApprovalVO);
     }
